move css file style folder an add style to navbar
