+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import './style/App.css';
 import {
   BrowserRouter as Router,
   Routes,
