+++ src/components/NavBar.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import '../style/navbar.css';
 
 const NavBar = () => (
   <div className="nav">
