+++ src/style/calculator.css
@@ -1,7 +1,3 @@
-.App {
-  text-align: center;
-}
-
 .calc {
   margin: auto;
   display: grid;
