+++ src/style/navbar.css
@@ -1,0 +1,22 @@
+li {
+  list-style: none;
+  padding: 1rem;
+}
+
+ul {
+  display: flex;
+  flex-direction: row;
+}
+
+.nav {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 80%;
+  margin: auto;
+  background-color: lightgray;
+}
+
+h1 {
+  margin: 1rem;
+}
